api/requirements.txt
Django==2.1.4
django-braces==1.13.0
django-debug-toolbar==1.11
django-extensions==2.1.4
django-filter==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.9.0
djangorestframework-csv==2.1.0
djangorestframework-xml==1.4.0
djangorestframework-gis==0.14
drf-extensions==0.4.0
drf-amsterdam==0.2.2
graypy==0.3.1
psycopg2-binary==2.7.6.1
factory-boy==2.11.1
drf-yasg==1.11.1
swagger-spec-validator==2.4.1

datapunt-authorization-django==0.2.23
datapunt-authorization-levels==0.1.3
tox==3.5.3
pytest==4.0.1
pytest-cov==2.6.0
pytest-flake8==1.0.2
pytest-faker==2.0.0
pytest-isort==0.2.1
pytest-env==0.6.2
pytest-mock==1.10.0
pytest-django==3.4.4
## The following requirements were added by pip freeze:
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
cryptography==2.3
defusedxml==0.5.0
Faker==1.0.0
filelock==3.0.10
flake8==3.6.0
idna==2.8
inflection==0.3.1
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
jsonschema==2.6.0
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==4.3.0
openapi-codec==1.3.2
pluggy==0.8.0
py==1.7.0
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
Pygments==2.3.0
PyJWT==1.7.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.20.1
ruamel.yaml==0.15.80
simplejson==3.16.0
six==1.11.0
sqlparse==0.2.4
text-unidecode==1.2
toml==0.10.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.1.0
